feat(containers-shared): Move `isDockerfile` utility function to `containers-shared`
