+++ packages/containers-shared/tests/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "@cloudflare/workers-tsconfig/tsconfig.json",
+	"compilerOptions": {
+		"module": "CommonJS",
+		"types": ["node", "vitest/globals"]
+	},
+	"include": ["**/*.ts"]
+}
